[REFACTOR] 파일 경로 수정
- Account/Table/... => Account/...
